refactor: split channel settings import/export
