+++ src/components/ViewController/ChannelsSettingsSection/ChannelControllers/ChannelSettingsImportExportButtons/ChannelSettingsImportExportButtons.tsx
@@ -1,290 +1,16 @@
 import { Box, Button, SxProps, Theme, Tooltip, alpha, useTheme } from '@mui/material';
-import { useCallback } from 'react';
-import { useDropzone } from 'react-dropzone';
-import { useSnackbar } from 'notistack';
 import DownloadIcon from '@mui/icons-material/Download';
 import UploadIcon from '@mui/icons-material/Upload';
-import { ChannelsSettings, PropertiesUpdateType, useChannelsStore } from '../../../../../stores/ChannelsStore';
-import { useViewerStore } from '../../../../../stores/ViewerStore';
-import { useShallow } from 'zustand/react/shallow';
 import { useTranslation } from 'react-i18next';
-
-const validateChannelImportData = (data: unknown): void => {
-  if (!data || typeof data !== 'object' || Array.isArray(data)) {
-    throw new Error('Root must be an object');
-  }
-  const obj = data as Record<string, unknown>;
-  if (!Array.isArray(obj.channels)) {
-    throw new Error('Missing or invalid "channels" array');
-  }
-  for (const ch of obj.channels) {
-    if (!ch || typeof ch !== 'object' || typeof (ch as Record<string, unknown>).name !== 'string') {
-      throw new Error('Each channel must have a string "name"');
-    }
-  }
-};
+import { useChannelSettingsImportExport } from './useChannelSettingsImportExport.hook';
 
 export const ChannelSettingsImportExportButtons = () => {
   const theme = useTheme();
   const sx = styles(theme);
-  const { enqueueSnackbar } = useSnackbar();
   const { t } = useTranslation();
 
-  const [ids, selections, channelsVisible, colors, contrastLimits, channelsSettings, setPropertiesForChannel] =
-    useChannelsStore(
-      useShallow((store) => [
-        store.ids,
-        store.selections,
-        store.channelsVisible,
-        store.colors,
-        store.contrastLimits,
-        store.channelsSettings,
-        store.setPropertiesForChannel
-      ])
-    );
-
-  const [channelOptions, metadata, generalDetails] = useViewerStore(
-    useShallow((store) => [store.channelOptions, store.metadata, store.generalDetails])
-  );
-
-  const exportChannelSettings = () => {
-    try {
-      const exportData = {
-        channels: ids.map((_, index) => {
-          const name = channelOptions[(selections as any)[index].c];
-          const perChannelSettings = channelsSettings?.[name];
-          return {
-            name,
-            visible: channelsVisible[index],
-            color: colors[index],
-            contrastLimits: contrastLimits[index],
-            ...(perChannelSettings?.initialContrastLimits !== undefined && {
-              initialContrastLimits: perChannelSettings.initialContrastLimits
-            }),
-            selection: selections[index]
-          };
-        })
-      };
-
-      let jsonData;
-      try {
-        jsonData = JSON.stringify(exportData, null, 2);
-      } catch (jsonError) {
-        console.error('Error stringifying channel settings:', jsonError);
-        enqueueSnackbar({
-          message: t('channelSettings.channelFormattingError'),
-          variant: 'error'
-        });
-        return;
-      }
-
-      const blob = new Blob([jsonData], { type: 'application/json' });
-      const url = URL.createObjectURL(blob);
-
-      let fileName = 'channel-settings';
-
-      if (generalDetails && generalDetails.fileName) {
-        const baseName = generalDetails.fileName.split('.').slice(0, -1).join('.');
-        const sanitizedName = baseName.replace(/[^\w-]/g, '_');
-        if (sanitizedName) {
-          fileName = `channel-settings_${sanitizedName}`;
-        }
-      } else if (metadata && metadata.Name) {
-        const baseName = metadata.Name.split('.').slice(0, -1).join('.');
-        const sanitizedName = baseName.replace(/[^\w-]/g, '_');
-        if (sanitizedName) {
-          fileName = `channel-settings_${sanitizedName}`;
-        }
-      }
-
-      const a = document.createElement('a');
-      a.href = url;
-      a.download = `${fileName}.json`;
-      document.body.appendChild(a);
-      a.click();
-
-      document.body.removeChild(a);
-      URL.revokeObjectURL(url);
-
-      enqueueSnackbar({
-        message: t('channelSettings.channelExportSuccess'),
-        variant: 'success'
-      });
-    } catch (error) {
-      console.error('Error exporting channel settings:', error);
-      enqueueSnackbar({
-        message: t('channelSettings.channelExportError'),
-        variant: 'error'
-      });
-    }
-  };
-
-  const onDrop = useCallback(
-    (acceptedFiles: File[]) => {
-      if (acceptedFiles.length === 0) return;
-
-      const file = acceptedFiles[0];
-      const reader = new FileReader();
-
-      reader.onload = (e) => {
-        const content = e.target?.result as string;
-        let importData: unknown;
-        try {
-          importData = JSON.parse(content);
-        } catch {
-          enqueueSnackbar({ message: t('channelSettings.channelImportError'), variant: 'error' });
-          return;
-        }
-
-        try {
-          validateChannelImportData(importData);
-        } catch {
-          enqueueSnackbar({ message: t('channelSettings.channelImportInvalidFormat'), variant: 'error' });
-          return;
-        }
-
-        try {
-          const data = importData as { channels: any[]; channelsSettings?: Record<string, any> };
-          const { addIsChannelLoading, setIsChannelLoading } = useViewerStore.getState();
-
-          const newChannelsSettings: ChannelsSettings = {};
-          data.channels.forEach((ch: any) => {
-            if (ch.name) {
-              newChannelsSettings[ch.name] = {
-                ...(ch.color !== undefined && { color: ch.color }),
-                ...(ch.initialContrastLimits !== undefined && { initialContrastLimits: ch.initialContrastLimits }),
-                ...(ch.contrastLimits !== undefined && {
-                  minValue: ch.contrastLimits[0],
-                  maxValue: ch.contrastLimits[1]
-                }),
-              };
-            }
-          });
-
-          const mergedSettings: ChannelsSettings = { ...newChannelsSettings, ...(data.channelsSettings ?? {}) };
-          if (Object.keys(mergedSettings).length > 0) {
-            useChannelsStore.setState({ channelsSettings: mergedSettings });
-          }
-
-          data.channels.forEach((channelData: any, idx: number) => {
-            if (idx < ids.length && channelData.name) {
-              const channelName = channelData.name;
-              const channelIndex = channelOptions.indexOf(channelName);
-
-              if (channelIndex !== -1) {
-                const newProps: Partial<PropertiesUpdateType> = {};
-
-                if (channelData.contrastLimits) {
-                  newProps.contrastLimits = channelData.contrastLimits;
-                }
-
-                if (channelData.color) {
-                  newProps.colors = channelData.color;
-                }
-
-                if (channelData.visible !== undefined) {
-                  newProps.channelsVisible = channelData.visible;
-                }
-
-                if (channelData.selection) {
-                  const updatedSelection = {
-                    ...channelData.selection,
-                    c: channelIndex
-                  };
-                  newProps.selections = updatedSelection;
-                }
-
-                setPropertiesForChannel(idx, newProps);
-              }
-            }
-          });
-
-          if (ids.length > data.channels.length) {
-            const { removeIsChannelLoading } = useViewerStore.getState();
-            for (let i = ids.length - 1; i >= data.channels.length; i--) {
-              useChannelsStore.getState().removeChannel(i);
-              removeIsChannelLoading(i);
-            }
-          }
-
-          if (data.channels.length > ids.length) {
-            const channelsToAdd = data.channels.slice(ids.length);
-            const channelIndicesToLoad: number[] = [];
-
-            channelsToAdd.forEach((channelData: any) => {
-              if (channelData.name) {
-                const channelName = channelData.name;
-                const channelIndex = channelOptions.indexOf(channelName);
-
-                if (channelIndex !== -1) {
-                  const newSelection = channelData.selection
-                    ? {
-                        ...channelData.selection,
-                        c: channelIndex
-                      }
-                    : {
-                        z: 0,
-                        c: channelIndex,
-                        t: 0
-                      };
-
-                  const numSelectionsBeforeAdd = useChannelsStore.getState().selections.length;
-                  channelIndicesToLoad.push(numSelectionsBeforeAdd);
-
-                  addIsChannelLoading(true);
-                  useChannelsStore.getState().addChannel({
-                    selections: newSelection,
-                    ids: String(Math.random()),
-                    channelsVisible: channelData.visible !== undefined ? channelData.visible : true,
-                    colors: channelData.color || [255, 255, 255],
-                    contrastLimits: (channelData.contrastLimits ?? [0, 65535]) as [number, number],
-                    domains: (channelData.contrastLimits ?? [0, 65535]) as [number, number]
-                  } as any);
-                }
-              }
-            });
-
-            if (channelIndicesToLoad.length > 0) {
-              let loadedCount = 0;
-              useViewerStore.setState({
-                onViewportLoad: () => {
-                  if (loadedCount < channelIndicesToLoad.length) {
-                    setIsChannelLoading(channelIndicesToLoad[loadedCount], false);
-                    loadedCount++;
-                    if (loadedCount === channelIndicesToLoad.length) {
-                      useViewerStore.setState({ onViewportLoad: () => {} });
-                    }
-                  }
-                }
-              });
-            }
-          }
-
-          enqueueSnackbar({
-            message: t('channelSettings.channelImportSuccess'),
-            variant: 'success'
-          });
-        } catch (error) {
-          console.error('Error importing channel settings:', error);
-          enqueueSnackbar({
-            message: t('channelSettings.channelImportError'),
-            variant: 'error'
-          });
-        }
-      };
-
-      reader.readAsText(file);
-    },
-    [ids, channelOptions, setPropertiesForChannel, enqueueSnackbar, t]
-  );
-
-  const { getRootProps, getInputProps, isDragActive, isDragAccept, isDragReject } = useDropzone({
-    onDrop,
-    accept: {
-      'application/json': ['.json']
-    },
-    maxFiles: 1
-  });
+  const { exportChannelSettings, getRootProps, getInputProps, isDragActive, isDragAccept, isDragReject } =
+    useChannelSettingsImportExport();
 
   let dynamicButtonText = t('general.import');
